requirements.txt
Flask==2.3.3
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
google-generativeai==0.3.0
tensorflow==2.13.0
Pillow==10.0.1
werkzeug==2.3.7
numpy==1.24.3
